+++ lib/kanayago/scope_node.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Kanayago
+  class ScopeNode
+    attr_reader :args, :body
+  end
+
+  class BlockNode # rubocop:disable Lint/EmptyClass
+  end
+
+  class ArgumentsNode
+    attr_reader :ainfo
+  end
+end