+++ lib/kanayago/variable_node.rb
@@ -20,4 +20,8 @@ module Kanayago
   class GlobalVariableNode
     attr_reader :vid
   end
+
+  class ConstantNode
+    attr_reader :vid
+  end
 end